navpointsAPF implemented
navpoints have been added which determines the negative portion of the apf.
TODO:
properly normalize image to displlay negative values

delete the nave point apf once the ego vehicle is reached.
Maybe add a "active" term in the navpoint dict to see which navpoint should be part of the apf, and hide all others.